requirements.txt
matplotlib==3.10.7
mysql_connector_repackaged==0.3.1
Pillow==12.0.0
python-dotenv==1.2.1
tkinterdnd2==0.4.3
ttkbootstrap==1.14.7
